Move typed ConfigEntry to coordinator module in point
