+++ homeassistant/components/point/coordinator.py
@@ -3,29 +3,29 @@
 from collections.abc import Callable
 from datetime import datetime
 import logging
-from typing import TYPE_CHECKING, Any
+from typing import Any
 
 from pypoint import PointSession
 
+from homeassistant.config_entries import ConfigEntry
 from homeassistant.core import HomeAssistant
 from homeassistant.helpers.update_coordinator import DataUpdateCoordinator, UpdateFailed
 from homeassistant.util.dt import parse_datetime
 
 from .const import DOMAIN, SCAN_INTERVAL
 
-if TYPE_CHECKING:
-    from . import PointConfigEntry
-
 _LOGGER = logging.getLogger(__name__)
+
+type PointConfigEntry = ConfigEntry[PointDataUpdateCoordinator]
 
 
 class PointDataUpdateCoordinator(DataUpdateCoordinator[dict[str, dict[str, Any]]]):
     """Class to manage fetching Point data from the API."""
 
-    config_entry: "PointConfigEntry"
+    config_entry: PointConfigEntry
 
     def __init__(
-        self, hass: HomeAssistant, point: PointSession, config_entry: "PointConfigEntry"
+        self, hass: HomeAssistant, point: PointSession, config_entry: PointConfigEntry
     ) -> None:
         """Initialize."""
         super().__init__(
